wip: add url groups

+++ src/db/models/url_group.go
@@ -6,7 +6,7 @@ type UrlGroup struct {
 	gorm.Model
 	ID        uint64 `gorm:"primaryKey;autoIncrement:false"`
 	Name      string `gorm:"unique,not null,size:10"`
-	CreatorID uint   `gorm:"not null"`
+	CreatorID uint64 `gorm:"not null"`
 	Creator   User   `gorm:"foreignKey:CreatorID"`
 }
